+++ terkin/datalogger.py
@@ -123,9 +123,10 @@ class TerkinDatalogger:
         # Bootstrap infrastructure.
         self.device.start_networking()
 
-        # Conditionally start telemetry if networking is available.
+        # Conditionally start network services and telemetry if networking is available.
         if self.device.status.networking:
             self.device.start_telemetry()
+            self.device.start_network_services()
 
         # Todo: Signal readyness by publishing information about the device (Microhomie).
         # e.g. ``self.device.publish_properties()``